Revamp README with structured release package table and shields
Updated the README to include a structured table of release packages for various Ezurio products, enhancing clarity and organization. 
This gets around the limit of 6 repo pins but still renders well and includes the additional info on repo watchers, forks and size that are available in the built in pin functionality.